feat: move poll list and create poll component file to a structured directory
